+++ lib/html_to/configuration.rb
@@ -2,7 +2,6 @@ module HtmlTo
   module Configuration
     class << self
       attr_accessor :chromium_path
-      attr_writer :global_templates_path
 
       def config
         yield self