+++ spec/spec_helper.rb
@@ -13,7 +13,7 @@ require 'sqlite3'
 require 'active_job'
 
 RSpec.configure do |config|
-
+  config.include ActiveJob::TestHelper
   config.before(:suite) do
     DatabaseCleaner.strategy = :transaction
     DatabaseCleaner.clean_with(:truncation)
@@ -23,6 +23,9 @@ RSpec.configure do |config|
     DatabaseCleaner.cleaning do
       example.run
     end
+  end
+  config.after(:each) do
+    FileUtils.rm_rf( Dir.glob(Rails.root.join("tmp/storage/*")) )
   end
 
 end
@@ -35,17 +38,10 @@ end
 Dir["#{File.dirname(__FILE__)}/support/**/*.rb"].sort.each { |f| require f }
 ENV['RAILS_ENV'] = 'test'
 require_relative '../spec/dummy/config/environment'
-ENV['RAILS_ROOT'] ||= "#{File.dirname(__FILE__)}../../../spec/dummy"
 
 TEST_DATABASE_FILE = File.expand_path('dummy/db/test.sqlite3', __dir__)
-FileUtils.rm( TEST_DATABASE_FILE ) rescue nil
+FileUtils.rm_f( TEST_DATABASE_FILE )
 ActiveRecord::Base.logger = Logger.new(STDOUT)
 ActiveRecord::Base.logger.level = Logger::WARN
-ActiveRecord::Base.connection.create_table :dummy_models do |t|
-  t.string :title
-  t.integer :description
-end
-ActiveRecord::Base.connection.create_table :dummy_classes do |t|
-  t.string :title
-  t.integer :description
-end
+ActiveRecord::MigrationContext.new("spec/dummy/db/migrate").migrate
+