+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # Changelog
-
-## 1.0.0 (12-May-2021)
-* 
+## 1.0.1
+Bug fixes. test refactor. Added integration tests
+## 1.0.0 (5-Sep-2023)
+Added configuration via DSL. Added ability to generate few images on one model. Replaced carrierwave uploader to active storage. Added rspec.
 ## 0.7.2 (12-May-2021)
 * Fixing Issues
 ## 0.7.0 (07-Apr-2021)